+++ company/.gitattributes
