+++ company/src/main/java/com/springcloud/company/company/repository/CompanyRepository.java
@@ -1,0 +1,14 @@
+package com.springcloud.company.company.repository;
+
+import com.springcloud.company.company.entity.Company;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface CompanyRepository extends JpaRepository<Company, UUID> {
+    Optional<Company> findByUserId(UUID userId);
+
+}
