+++ company/src/main/java/com/springcloud/company/product/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package com.springcloud.company.product.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
