+++ company/src/main/java/com/springcloud/company/product/config/PropertyConfig.java
@@ -1,0 +1,13 @@
+package com.springcloud.company.product.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.annotation.PropertySources;
+
+@Configuration
+@PropertySources({
+        @PropertySource("classpath:properties/env.properties") // env.properties 파일 소스 등록
+})
+public class PropertyConfig {
+
+}
