+++ company/src/main/java/com/springcloud/company/product/dto/ProductResponseDto.java
@@ -1,0 +1,46 @@
+package com.springcloud.company.product.dto;
+
+import com.springcloud.company.company.entity.Company;
+import com.springcloud.company.product.entity.Product;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductResponseDto {
+
+    //상품 ID
+    private UUID Id;
+    //허브 ID
+    private UUID hubId;
+    //유저ID
+    private UUID userId;
+    //업체ID
+    private Company company;
+    //상품명
+    private String productName;
+    //상품 가격
+    private Integer price;
+    //출고 가능 수량
+    private Integer stock;
+
+
+    public ProductResponseDto(Product product) {
+        this.Id = product.getId();
+        this.hubId = product.getHubId();
+        this.userId = product.getUserId();
+        this.company = product.getCompany();
+        this.productName = product.getProductName();
+        this.price = product.getPrice();
+        this.stock = product.getStock();
+
+    }
+
+
+}
