+++ company/src/main/java/com/springcloud/company/product/repository/ProductRepository.java
@@ -1,0 +1,11 @@
+package com.springcloud.company.product.repository;
+
+import com.springcloud.company.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, UUID> {
+}
