--- product/src/main/java/com/springcloud/product/ProductApplication.java
@@ -1,14 +1,0 @@
-package com.springcloud.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProductApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProductApplication.class, args);
-	}
-
-
-}
