--- product/src/main/java/com/springcloud/product/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.springcloud.product.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
