--- product/src/main/java/com/springcloud/product/config/PropertyConfig.java
@@ -1,13 +1,0 @@
-package com.springcloud.product.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.context.annotation.PropertySources;
-
-@Configuration
-@PropertySources({
-        @PropertySource("classpath:properties/env.properties") // env.properties 파일 소스 등록
-})
-public class PropertyConfig {
-
-}
