--- product/src/main/java/com/springcloud/product/dto/ProductRequestDto.java
@@ -1,15 +1,0 @@
-package com.springcloud.product.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ProductRequestDto {
-    //상품명
-    private String productName;
-    //상품 가격
-    private Integer price;
-    //출고 가능 수량
-    private Integer stock;
-}
