--- product/src/main/java/com/springcloud/product/entity/BaseEntity.java
@@ -1,55 +1,0 @@
-package com.springcloud.product.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.Comment;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-    @CreatedDate
-    @Column(name = "created_at", updatable = false)
-    @Temporal(TemporalType.TIMESTAMP)
-    @Comment("레코드 생성 일시")
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(name = "created_by", length = 50, updatable = false)
-    @Comment("레코드 생성자")
-    private String createdBy;
-
-    @LastModifiedDate
-    @Column(name = "updated_at")
-    @Temporal(TemporalType.TIMESTAMP)
-    @Comment("레코드 수정 일시")
-    private LocalDateTime updatedAt;
-
-    @LastModifiedBy
-    @Column(name = "updated_by", length = 50)
-    @Comment("레코드 수정자")
-    private String updatedBy;
-
-    @Column(name = "deleted_at")
-    @Temporal(TemporalType.TIMESTAMP)
-    @Comment("레코드 삭제 일시")
-    private LocalDateTime deletedAt;
-
-    @Column(name = "deleted_by", length = 50)
-    @Comment("레코드 삭제자")
-    private String deletedBy;
-
-    public void delete(String deletedBy) {
-        this.deletedBy = deletedBy;
-        this.deletedAt = LocalDateTime.now();
-    }
-}
