--- product/src/main/java/com/springcloud/product/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.springcloud.product.repository;
-
-import com.springcloud.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, UUID> {
-}
